# docs/spec/api/openapi.yaml
openapi: 3.1.0
info:
  title: IEapp API
  version: "0.1"
  description: |
    Minimal OpenAPI document for the current REST surface.
servers:
  - url: http://localhost:8000
paths:
...
  /api/workspaces/{workspace_id}/classes:
    get:
      summary: List classes
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    post:
      summary: Upsert class
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Created
  /api/workspaces/{workspace_id}/classes/types:
    get:
      summary: List class column types
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /api/workspaces/{workspace_id}/attachments:
    get:
      summary: List attachments
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    post:
      summary: Upload attachment
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Created
  /api/workspaces/{workspace_id}/attachments/{attachment_id}:
    delete:
      summary: Delete attachment
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
        - name: attachment_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /api/workspaces/{workspace_id}/links:
    get:
      summary: List links
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    post:
      summary: Create link
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Created
  /api/workspaces/{workspace_id}/links/{link_id}:
    delete:
      summary: Delete link
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
        - name: link_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK